Фикс вылета на запуске

+++ Byster/Byster/Views/LoadingWindow.xaml.cs
@@ -100,19 +100,22 @@ namespace Byster.Views
                 BackgroundImageDownloader.Init();
                 LogInfo("Common", "Запущен BackgroundImageDownloader");
                 LogInfo("Common", "Удаление остаточных файлов");
-                if (File.Exists("BysterUpdate.exe")) File.Delete("BysterUpdate.exe");
-                if (File.Exists("update.bat")) File.Delete("update.bat");
-                if (File.Exists("changeLocalization.bat")) File.Delete("changeLocalization.bat");
+                if (File.Exists("BysterUpdate.exe")) try { File.Delete("BysterUpdate.exe"); } catch { LogWarn("Common", "Ошибка удаления BysterUpdate.exe"); } catch { LogError("Common", "Ошибка удаления changeLocalization.bat"); }
+                if (File.Exists("update.bat")) try { File.Delete("update.bat"); } catch { LogWarn("Common", "Ошибка удаления update.bat"); }
+                if (File.Exists("changeLocalization.bat")) try { File.Delete("changeLocalization.bat"); } catch { LogWarn("Common", "Ошибка удаления changeLocalization.bat"); }
                 string qrCodesPath = Path.Combine(Path.GetTempPath(), "BysterQRCodes");
-                foreach(var file in Directory.GetFiles(qrCodesPath))
+                if(Directory.Exists(qrCodesPath))
                 {
-                    try
+                    foreach (var file in Directory.GetFiles(qrCodesPath))
                     {
-                        File.Delete(Path.Combine(qrCodesPath, file));
-                    }
-                    catch
-                    {
-                        LogWarn("Common", "Ошибка удаления QR Code", "Path: ", file);
+                        try
+                        {
+                            File.Delete(Path.Combine(qrCodesPath, file));
+                        }
+                        catch
+                        {
+                            LogWarn("Common", "Ошибка удаления QR Code", "Path: ", file);
+                        }
                     }
                 }
                 LogInfo("Common", "Удаление остаточных файлов завершено");
